Rename packages for clarity
Originally the packages were named as 'create_autonomy', with the 'ca_' suffix for short.
This was due to a name conflict with the turtlebot packages (e.g. create_driver and create_description):

https://github.com/turtlebot/turtlebot_create.git

Since the turtlebot packages are not released into ROS Melodic, the plan is to release these packages instead.
If the turtlebot packages are eventually released, it should be straightforward to adapt them to use the
description and driver packages here.

Summary of renaming:

* create_autonomy -> create_robot
* ca_description -> create_description
* ca_driver -> create_driver
* ca_msgs -> create_msgs
* ca_tools -> create_bringup

Signed-off-by: Jacob Perron <jacob@openrobotics.org>